% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{compute_mu_hat}
\alias{compute_mu_hat}
\title{Estimates mu_hat for a tree}
\usage{
compute_mu_hat(tree, nperm)
}
\arguments{
\item{tree}{A \code{phylo} object: a ranked, planar, partially labeled tree shape}

\item{nperm}{Integer: number of planar permutations of the given tree}
}
\value{
An integer with the value of mu_hat
}
\description{
This function estimates the mean number of attachments for a given tree shape
with a set number of permutations
}
