% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{compute_S_mean}
\alias{compute_S_mean}
\title{Estimates S_mean for a tree}
\usage{
compute_S_mean(tree, nperm)
}
\arguments{
\item{tree}{A \code{phylo} object: a ranked, planar, partially labeled tree shape}

\item{nperm}{Integer: number of planar permutations of the given tree}
}
\value{
An integer with the value of S_mean
}
\description{
This function estimates the mean number of attachments for a given tree shape
with a set number of permutations
}
