% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{count_partial_labels}
\alias{count_partial_labels}
\title{Count the number of partial labelings on a given tree shape}
\usage{
count_partial_labels(tree, B)
}
\arguments{
\item{tree}{A ranked tree shape}

\item{B}{Number of blue tips}
}
\value{
The number of possible partial labelings
}
\description{
This function counts the number of possible partial labelings with N tips and B blue
on the given ranked tree shape
}
