% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{count_ranked_pl_trees}
\alias{count_ranked_pl_trees}
\title{Count the number of ranked partially labeled tree shapes with N tips and B blue}
\usage{
count_ranked_pl_trees(N, B)
}
\arguments{
\item{N}{Number of tips}

\item{B}{Number of blue tips}
}
\value{
The number of ranked partially labeled tree shapes with N tips and B blue
}
\description{
Count the number of ranked partially labeled tree shapes with N tips and B blue
}
