% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_tree.R
\name{count_same_attachments}
\alias{count_same_attachments}
\title{This function computes the number of same attachments for the given tree}
\usage{
count_same_attachments(tree, nodes_to_switch = c(), tip_order = FALSE)
}
\arguments{
\item{tree}{A \code{phylo} object: a ranked, planar, partially labeled tree shape}

\item{tip_order}{Boolean for whether to return the order the tips were added or not}

\item{nodes_to_switch:}{vector of internal nodes to switch the left and right subtrees when calculating S (default empty)}
}
\value{
Integer: number of same attachments
}
\description{
This function computes the number of same attachments for the given tree
}
