% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_representation_conversions.R
\name{create_tree_attachments}
\alias{create_tree_attachments}
\title{Construct tree from a sequence of attachments and type of attachments}
\usage{
create_tree_attachments(attachment_matrix)
}
\arguments{
\item{attachment_matrix}{A matrix with three columns: attaching (int), attached_to (int), and same (boolean)}
}
\value{
A ranked, planar, partially labeled tree
}
\description{
This function constructs a tree object based on the given attachment matrix
}
