% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_representation_conversions.R
\name{crp_tables_to_colors}
\alias{crp_tables_to_colors}
\title{Find the attachments and color type for a list of tables}
\usage{
crp_tables_to_colors(tables_list)
}
\arguments{
\item{tables_list}{A list of vectors indicating the list of tables}
}
\value{
A matrix with three columns: attaching (int), attached_to (int), and same (boolean). Again: rightmost tip = 1, leftmost tip = N
}
\description{
This function computes the attachments and type of attachment (same color or not) for a list of tables satisfying rules to correspond to a CRP-TREE tree
}
