% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{crp_tree_power}
\alias{crp_tree_power}
\title{Compute the power of the testing methods conditional on a given tree shape}
\usage{
crp_tree_power(
  tree,
  alphas,
  sig_level = 0.05,
  num_sample = 500,
  num_pl = 500,
  nperm_null = 300,
  nperm_labels = 300,
  acceptances = FALSE
)
}
\arguments{
\item{tree}{A phylo object: ranked and partially labeled with two categories}

\item{alphas}{Vector of alphas for which to compute the power under}

\item{sig_level}{Numeric value of the significance level}

\item{num_sample}{Integer: number of iterations to run the labeling chain of the MCMC}

\item{num_pl}{Integer: number of iterations to tun the planar chain of the MCMC}

\item{nperm_null}{Integer: number of planar permutations of the given tree}

\item{nperm_labels}{Integer: number of label permutations of the given tree}

\item{acceptances}{Boolean whether to return the number of acceptances as well (default \code{FALSE})}
}
\value{
A matrix with the power of each method under each alpha (and possible the number of acceptances)
}
\description{
This function computes the power of four testing methods (pval_tree, pval_avg, PS, AI) by
\itemize{
\item sampling planar representations of the given tree using MCMC under specific alpha values
\item computing the null thresholds
Also can return the number of acceptances in each step of the MCMC
}
}
