% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_yfv}
\alias{ex_yfv}
\title{Yellow Fever Virus posterior trees}
\format{
\subsection{\code{ex_yfv}}{

A Nexus file with 100 posterior trees sampled from Yellow Fever Virus sequences
}
}
\source{
\url{https://beast.community/workshop_continuous_diffusion_yfv}
}
\usage{
ex_yfv
}
\description{
Trees generated from BEAST using 65 Yellow Fever Virus sequences
}
\keyword{datasets}
