% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{expected_value_function}
\alias{expected_value_function}
\title{Compute expected value of S for a tree generated under the CRP-TREE model}
\usage{
expected_value_function(N, B, alpha)
}
\arguments{
\item{N}{Number of tips}

\item{B}{Number of blue tips}

\item{alpha}{Numeric parameter for the tree model}
}
\value{
The expected value
}
\description{
This function computes the expected value of S for a tree generated under CRP-TREE(N, B, alpha)
}
