% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_representation_conversions.R
\name{get_crp_tables}
\alias{get_crp_tables}
\title{Find the "tables" associated with a CRP-TREE tree}
\usage{
get_crp_tables(tree)
}
\arguments{
\item{tree}{A phylo object: a ranked, planar, partially labeled tree shape}
}
\value{
List of tables where the numbers corresponds to the tips on the tree with rightmost = 1, leftmost = N
}
\description{
This function computes the table representation associated with a CRP-TREE tree
}
