% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{get_mle_1tree}
\alias{get_mle_1tree}
\title{Compute the MLE of alpha for a ranked, planar, partially labeled tree}
\usage{
get_mle_1tree(tree)
}
\arguments{
\item{tree}{A ranked, planar, partially labeled tree}
}
\value{
A vector with the MLE and maximum log-likelihood value
}
\description{
This function computes the MLE of alpha for a ranked, planar, partially labeled tree
generated under CRP-TREE(N,B, alpha)
}
