% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{get_mle_trees}
\alias{get_mle_trees}
\title{Computes the MLE of alpha for a list of ranked, planar, partially labeled trees}
\usage{
get_mle_trees(tree_list)
}
\arguments{
\item{tree_list}{A list of ranked, planar, partially labeled trees}
}
\value{
A vector with the MLE and maximum log-likelihood value
}
\description{
This function computes the MLE of alpha for a list of ranked, planar, partially labeled trees
with the same N generated under CRP-TREE(N,B, alpha)
}
