% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{lrt_1tree}
\alias{lrt_1tree}
\title{Applies LRT for a ranked, planar, partially labeled tree}
\usage{
lrt_1tree(tree)
}
\arguments{
\item{tree}{A ranked, planar, partially labeled tree}
}
\value{
A vector with the MLE and p-value
}
\description{
This function applies the LRT for testing alpha = 1 vs alpha > 1 for a
ranked, planar, partially labeled tree generated under CRP-TREE(N,B, alpha)
}
