% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{lrt_many_trees}
\alias{lrt_many_trees}
\title{Applies LRT for a list of ranked, planar, partially labeled trees}
\usage{
lrt_many_trees(tree_list)
}
\arguments{
\item{tree}{A ranked, planar, partially labeled tree}
}
\value{
A vector with the MLE and p-value
}
\description{
This function applies the LRT for testing alpha = 1 vs alpha > 1 for a list of
ranked, planar, partially labeled trees generated under CRP-TREE(N,B, alpha)
with the same N and alpha
}
