% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_tree.R
\name{pcrp_tree}
\alias{pcrp_tree}
\title{Calculates the probability of a ranked, planar, partially labeled tree shape}
\usage{
pcrp_tree(tree, alpha, only_prob = TRUE)
}
\arguments{
\item{tree}{A \code{phylo} object: a ranked, planar, partially labeled tree shape}

\item{alpha}{Numeric value for alpha}

\item{only_prob}{Boolean specifying whether to return only the probability (default \code{TRUE}). If \code{FALSE}, returns a list with the log probability and sufficient statistics S and Ws}
}
\value{
Numeric of the probability or list of log probability and sufficient statistics
}
\description{
This function computes the probability of the ranked, planar, partially labeled
tree shape under the CRP-TREE model for a specific alpha value
}
