% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_tree.R
\name{plot_crp_tree}
\alias{plot_crp_tree}
\title{Plot a tree with two categories of labelings}
\usage{
plot_crp_tree(
  tree,
  node_labels = FALSE,
  cex_tip = 1.5,
  cex_node = 0.5,
  frame = "circle",
  bg = "darkseagreen1",
  align_tips = TRUE
)
}
\arguments{
\item{tree}{A phylo object to plot}

\item{node_labels}{Boolean on whether or not to highlight internal node labels}

\item{cex_tip}{Numeric for the scaling factor of the tip labels}

\item{cex_node}{Numeric for the scaling factor of the internal node labels}

\item{frame}{String specifying the shape of the internal node labels}

\item{bg}{String specifying the background color of internal node labels}

\item{align_tips}{Boolean specifying whether tips should be same length from root}
}
\description{
This function plots a partially labeled tree with tip colors red and blue
}
