% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_tree.R
\name{pnull_tree}
\alias{pnull_tree}
\title{This function computes the conditional probability of partially labeled tree shape}
\usage{
pnull_tree(tree)
}
\arguments{
\item{tree}{A phylo object: a ranked, partially labeled tree shape}
}
\value{
Numeric of the probability
}
\description{
This function computes the conditional probability of partially labeled tree shape
}
