% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{posterior_trees_bats}
\alias{posterior_trees_bats}
\title{Applies BaTS on a posterior distribution of trees}
\usage{
posterior_trees_bats(
  tree_list,
  tree_term = "",
  tree_tip_correspondence = NULL,
  n_bats = 500,
  nperm_obs = 500,
  nCores = 8
)
}
\arguments{
\item{tree_list}{A list of trees, usually a posterior distribution of trees}

\item{tree_term}{String specifying all tip labels containing this term are one category, and the rest form the other category}

\item{tree_tip_correspondence}{Matrix with two columns: first column is the tip labels of the tree, second is the category}

\item{n_bats}{Integer: number of label permutations of the given tree}

\item{nperm_obs}{Integer: number of planar permutations of the given tree to get mu_hat}

\item{nCores}{Integer: number of parallel processes to run}
}
\value{
A vector with the median values of mu_hat across BaTS
}
\description{
This function computes mu_hat, pval_tree, and pval_avg for a list of trees
It also processes the trees into desired format given a term (only term)
}
