% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{posterior_trees_pval}
\alias{posterior_trees_pval}
\title{Computes p-values for a posterior distribution of trees}
\usage{
posterior_trees_pval(
  tree_list,
  tree_term,
  tree_tip_correspondence = NULL,
  nperm_planar = 500,
  nperm_labels = 499,
  nCores = 8
)
}
\arguments{
\item{tree_list}{A list of ranked, planar, partially labeled trees}

\item{tree_term}{String specifying all tip labels containing this term are one category, and the rest form the other category}

\item{tree_tip_correspondence}{Matrix with two columns: first column is the tip labels of the tree, second is the category}

\item{nperm_planar}{Integer: number of planar permutations of the given tree}

\item{nperm_labels}{Integer: number of label permutations of the given tree}

\item{nCores}{Integer: number of parallel processes to run}
}
\value{
A matrix with the values of S_mean, pval_tree, and pval_avg for each tree
}
\description{
This function computes S_mean, pval_tree, and pval_avg for a list of trees
It also processes the trees into desired format given a term or matrix indicating
the categories of tip labels
}
