% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_tree.R
\name{rcrp_tree}
\alias{rcrp_tree}
\title{Generate a tree according to the CRP-TREE model}
\usage{
rcrp_tree(N, B, alpha)
}
\arguments{
\item{N}{Number of tips}

\item{B}{Number of tips labeled blue}

\item{alpha}{parameter of the CRP-TREE model}
}
\value{
A \code{phylo} object generated from CRP-TREE(alpha)
}
\description{
This function generates a ranked, planar, partially labeled tree under
the CRP-TREE(alpha) model with N tips and B labeled blue
}
\examples{
rcrp_tree(10, 5, 2)
}
