% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{simulate_mu_hat_tree}
\alias{simulate_mu_hat_tree}
\title{Simulate null distribution of mu_hat}
\usage{
simulate_mu_hat_tree(tree, nperm_planar, nperm_labels)
}
\arguments{
\item{tree}{A \code{phylo} object: a ranked, planar, partially labeled tree shape}

\item{nperm_planar}{Integer: number of planar permutations of the given tree}

\item{nperm_labels}{Integer: number of label permutations of the given tree}
}
\value{
A vector with the mu_hat values
}
\description{
This function simulates the null distribution of mu_hat for a given tree
}
