% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{simulate_S_grid}
\alias{simulate_S_grid}
\title{Simulates S for a grid of planarities and labels}
\usage{
simulate_S_grid(tree, nperm_planar, nperm_labels)
}
\arguments{
\item{tree}{A phylo object: a ranked, planar, partially labeled tree shape}

\item{nperm_planar}{Integer: number of planar permutations of the given tree}

\item{nperm_labels}{Integer: number of label permutations of the given tree}
}
\value{
A matrix with the S values
}
\description{
This function samples S for a grid of planarities and labels where
one row is one planar representation and one column is a partial labeling
}
