% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{simulate_S_mean_tree}
\alias{simulate_S_mean_tree}
\title{Simulate null distribution of S_mean}
\usage{
simulate_S_mean_tree(tree, nperm_planar, nperm_labels)
}
\arguments{
\item{tree}{A phylo object: a ranked, planar, partially labeled tree shape}

\item{nperm_planar}{Integer: number of planar permutations of the given tree}

\item{nperm_labels}{Integer: number of label permutations of the given tree}
}
\value{
A vector with the S_mean values
}
\description{
This function simulates the null distribution of S_mean for a given tree
}
